requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
PyMySQL>=1.1.0
python-dotenv==1.0.0
pandas>=2.1.0
numpy>=1.26.0
Pillow>=10.0.0
